chore: validate.ts 파일 수정, 오타 수정
