+++ src/dto/notification.dto.ts
@@ -1,4 +1,4 @@
-import { IsNotEmpty, IsOptional, IsString, MinLength } from 'class-validator';
+import { IsNotEmpty, IsOptional, IsString } from 'class-validator';
 import { Notification } from '@prisma/client';
 
 export class NotificationDTO {
