+++ src/lib/index.ts
@@ -4,7 +4,6 @@ export * from './errors/ForbiddenError';
 export * from './errors/NotFoundError';
 export * from './errors/UnauthorizedError';
 
-export * from './validator';
 export * from './constants';
 export * from './enums';
 export * from './inversify.config';
